Use of experimental gemm for small matrices (maybe useful for half precision)

+++ tests/bsr.cpp
@@ -981,7 +981,7 @@ int main(int argc, char **argv) {
         for (int i = 0; i < ncomponents; ++i) ctx.push_back(createCpuContext());
         std::vector<Cpu> xpus;
         for (const auto &i : ctx) xpus.push_back(i.toCpu(0));
-        test<std::complex<double>, Cpu>(dim, procs, rank, nprocs, max_power, nrep, ctx, xpus);
+        test<std::complex<float>, Cpu>(dim, procs, rank, nprocs, max_power, nrep, ctx, xpus);
         clearCaches();
     }
 #ifdef SUPERBBLAS_USE_GPU
